requirements.txt
pandas
tensorboard
fairscale
sentencepiece
gradio==4.19.0
packaging
transformers==4.46.2
diffusers==0.34.0
pyyaml
pathlib
Ninja
bitsandbytes
httpx[socks]
einops
regex
h5py
accelerate
pre-commit
torchao==0.11.0
pytorch_lightning
